Implement a module to use IBM speech to text API
In this module we use IBM speech to text API to get the text for all the filtered audios we generated with cut.py module. We then generate a dictionary with keys being the text data and values being relative addresses of audio files in order to access them later.